+++ src/jsMain/kotlin/dev/inmo/jsuikit/elements/List.kt
@@ -2,14 +2,32 @@ package dev.inmo.jsuikit.elements
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import dev.inmo.jsuikit.modifiers.UIKitModifier
-import dev.inmo.jsuikit.modifiers.include
+import dev.inmo.jsuikit.modifiers.*
+import dev.inmo.jsuikit.utils.Attrs
 import org.jetbrains.compose.web.dom.*
 import org.w3c.dom.HTMLHeadingElement
 import org.w3c.dom.HTMLUListElement
 
 @Composable
 fun <T> List(
+    data: SnapshotStateList<T>,
+    ukAttrs: Attrs<HTMLUListElement> = Attrs.empty(),
+    elementAllocator: @Composable ElementScope<HTMLUListElement>.(T) -> Unit
+) {
+    Ul(
+        {
+            include(UIKitList)
+            ukAttrs.builder(this)
+        }
+    ) {
+        data.forEach {
+            elementAllocator(it)
+        }
+    }
+}
+
+@Composable
+fun <T> ListWithTitle(
     title: String,
     data: SnapshotStateList<T>,
     vararg titleModifiers: UIKitModifier,
@@ -23,15 +41,27 @@ fun <T> List(
         Text(title)
     }
     besidesTitleAndList ?.invoke()
-    Ul(
-        {
-            classes("uk-list")
-            include(*ulModifiers)
-            ulCustomizer()
-        }
-    ) {
-        data.forEach {
-            elementAllocator(it)
-        }
-    }
+    List(data, Attrs(*ulModifiers) { ulCustomizer(this) }, elementAllocator)
 }
+
+@Deprecated("Renamed", ReplaceWith("ListWithTitle", "dev.inmo.jsuikit.elements.ListWithTitle"))
+@Composable
+fun <T> List(
+    title: String,
+    data: SnapshotStateList<T>,
+    vararg titleModifiers: UIKitModifier,
+    ulModifiers: Array<UIKitModifier> = emptyArray(),
+    besidesTitleAndList: (@Composable () -> Unit)? = null,
+    titleCustomizer: AttrBuilderContext<HTMLHeadingElement> = {},
+    ulCustomizer: AttrBuilderContext<HTMLUListElement> = {},
+    elementAllocator: @Composable ElementScope<HTMLUListElement>.(T) -> Unit
+) = ListWithTitle(
+    title,
+    data,
+    *titleModifiers,
+    ulModifiers = ulModifiers,
+    besidesTitleAndList = besidesTitleAndList,
+    titleCustomizer = titleCustomizer,
+    ulCustomizer = ulCustomizer,
+    elementAllocator = elementAllocator
+)