+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.53
 
+Improvements in `UIKitPadding`
+
 ## 0.0.52
 
 * Improve work with UIKitModifiers